Move new methods to devolved classes with unit tests
