+++ src/main/java/uk/gov/hmcts/reform/preapi/services/edit/EditRequestCrudService.java
@@ -18,12 +18,16 @@ import uk.gov.hmcts.reform.preapi.enums.EditRequestStatus;
 import uk.gov.hmcts.reform.preapi.enums.UpsertResult;
 import uk.gov.hmcts.reform.preapi.exception.BadRequestException;
 import uk.gov.hmcts.reform.preapi.exception.NotFoundException;
+import uk.gov.hmcts.reform.preapi.media.edit.EditInstructions;
 import uk.gov.hmcts.reform.preapi.repositories.EditRequestRepository;
 
 import java.sql.Timestamp;
 import java.time.Instant;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.Set;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Service
@@ -113,6 +117,23 @@ public class EditRequestCrudService {
             return Pair.of(UpsertResult.UPDATED, request);
         }
         return Pair.of(UpsertResult.CREATED, request);
+    }
+
+    @Transactional
+    public Set<UUID> findRecordingIdsWithForceReencodeRequests(Set<UUID> sourceRecordingIds) {
+        if (sourceRecordingIds.isEmpty()) {
+            return Set.of();
+        }
+
+        return editRequestRepository.findAllBySourceRecordingIdIn(sourceRecordingIds).stream()
+            .filter(editRequest -> {
+                EditInstructions instructions = EditInstructions.tryFromJson(editRequest.getEditInstruction());
+                return instructions != null && instructions.isForceReencode();
+            })
+            .map(EditRequest::getSourceRecording)
+            .filter(Objects::nonNull)
+            .map(Recording::getId)
+            .collect(Collectors.toSet());
     }
 
     private void validateEditMode(CreateEditRequestDTO dto) {
